% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abr_baseclass.R
\docType{class}
\name{abr_baseclass}
\alias{abr_baseclass}
\alias{abr_ecoffinder-class}
\title{Classes of package \pkg{antibioticR} ...}
\description{
The classes define ...
}
\details{
The \code{abr_ecoffinder}: class represents a distribution component
fitted with the ECOFFinder method.
}
