% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abr_methods.R
\docType{methods}
\name{obs}
\alias{obs}
\alias{obs,abr_baseclass-method}
\alias{coef,abr_baseclass-method}
\alias{fit,abr_baseclass-method}
\alias{deviance,abr_baseclass-method}
\alias{residuals,abr_baseclass-method}
\alias{df.residual,abr_baseclass-method}
\alias{startpar,abr_ecoffinder-method}
\alias{models,abr_ecoffinder-method}
\alias{ecoff,abr_ecoffinder-method}
\title{Accessor Methods of Package \pkg{antibioticR}.}
\usage{
obs(object, ...)

\S4method{obs}{abr_baseclass}(object, ...)

\S4method{coef}{abr_baseclass}(object, ...)

\S4method{fit}{abr_baseclass}(object, ...)

\S4method{deviance}{abr_baseclass}(object, ...)

\S4method{residuals}{abr_baseclass}(object, ...)

\S4method{df.residual}{abr_baseclass}(object, ...)

\S4method{startpar}{abr_ecoffinder}(object, ...)

\S4method{models}{abr_ecoffinder}(object, ...)

\S4method{ecoff}{abr_ecoffinder}(object, ...)
}
\arguments{
\item{object}{name of a 'antibioticR' object.
parameters is returned and printed.}

\item{\dots}{other arguments passed to the methods.

Note: this is a still dysfunctional template.}
}
\description{
Functions to access data and results of fitted antibioticR objects:
\code{summary}, \code{obs} \code{coeff}, ...
}
