% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abr_quantile.R
\name{abr_quantile}
\alias{abr_quantile}
\title{Normal Quantiles of Wild Type Population Identified with Ecoffinder}
\usage{
abr_quantile(obj, q, ...)
}
\arguments{
\item{obj}{object of class \linkS4class{abr_ecoffinder}}

\item{q}{vector of quantiles}

\item{...}{other arguments passed to \code{\link{qnorm}}}
}
\value{
vector of quantiles
}
\description{
Applies \code{\link{qnorm}} to an object returned by \code{ecoffinder_nls}.
}
\details{
If log2 was used for MIC data,
  the results are back-transformed by 2^x.
}
\examples{


}
