% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abr_summary.R
\name{summary,abr_baseclass-method}
\alias{summary,abr_baseclass-method}
\alias{summary,abr_ecoffinder-method}
\title{Summary of an 'abr_ecoffinder' Object}
\usage{
\S4method{summary}{abr_baseclass}(object, ...)

\S4method{summary}{abr_ecoffinder}(object, ...)
}
\arguments{
\item{object}{abr_ecoffinder object}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Returns parameters and quantiles of the normal distribution fitted to
  the wild type subset of the data.
}
