% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECOFFinder.R
\name{ECOFFinder}
\alias{ECOFFinder}
\title{Launch antibioticR App ECOFFinder}
\usage{
ECOFFinder(...)
}
\arguments{
\item{...}{optional parameters passed to the app (not yet implemented)}
}
\value{
This function has no return value.
}
\description{
Launch a shiny app to estimate MIC concentrations interactively.
}
\details{
The app is still experimental.
}
\examples{

\dontrun{
ECOFFinder()
}

}
