% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoffinder_nls.R
\name{ecoffinder_nls}
\alias{ecoffinder_nls}
\title{Find Cutoff Value by Nonlinear Regression to Truncated Data}
\usage{
ecoffinder_nls(
  conc,
  count,
  startpar = "mode",
  concentrations = NA,
  log2 = TRUE,
  plot = TRUE
)
}
\arguments{
\item{conc}{concentration of the antibiotic}

\item{count}{raw frequency}

\item{startpar}{start parameters for the nonlinear search,
or "mode" resp. "mean" for an automatic determination}

\item{concentrations}{which concentrations are tested}

\item{log2}{logical determining if conc are log-transformed or not}

\item{plot}{logical, switch visualization on or off}
}
\value{
an object of class \code{\link{abr_ecoffinder-class}} containing the fitted
  parameters and statistics of the final and intermediate fits.
}
\description{
Identifies the wild-type sub-population by fitting a cumulative normal
distribution to subsets of MIC or ID data.
}
\details{
Start values for the nonlinear regression can be automatically
  determined with function \code{\link{ecoffinder_startpar}}.

  The default search interval starts one concentration level above the mode resp. mean.
}
\examples{

## raw data contain NA values
data(micdata)
plot(freq ~ log2(conc), data=micdata, type="h")

## discard NA values
measured <- na.omit(micdata)

## cumulative plot
plot(cumsum(freq) ~ log2(conc), data=measured, type="l")

x <- log2(measured$conc)
y <- measured$freq

## heuristic start values
pstart <- ecoffinder_startpar(x, y)
pstart

## nonlinear regression
p <- ecoffinder_nls(x, y, pstart)
summary(p)

}
\references{
Turnidge, J., Kahlmeter, G., Kronvall, G. (2006) Statistical characterization of
  bacterial wild-type MIC value distributions and the determination of
  epidemiological cut-off values. Clin Microbial Infect 12: 418-425
  \doi{10.1111/j.1469-0691.2006.01377.x}
}
\seealso{
\code{\link{ecoffinder_startpar}} for heuristic methods to guess start parameters\cr
  \code{\link{ECOFFinder}} for an interactice shiny app
}
