% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixParmDf2parmlist.R
\name{mixParmDf2parmlist}
\alias{mixParmDf2parmlist}
\title{Convert Parameters in data frame format to a paremeter list ...}
\usage{
mixParmDf2parmlist(df)
}
\arguments{
\item{df}{data frame with parameters}
}
\value{

}
\description{
Convert Parameters in data frame format to a paremeter list ...
}
\details{
deprecated !!!
}
