% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostics,mxMle-method}
\title{Diagnostics of Mixture Objects Fitted with 'mle2'}
\usage{
diagnostics(obj, ...)

\S4method{diagnostics}{mxMle}(obj, breaks = NULL, counts = NULL, ...)
}
\arguments{
\item{obj}{object of class \code{mxMle}}

\item{...}{reserved for future extensions}

\item{breaks}{class boundaries of the data}

\item{counts}{frequency of observations}
}
\value{
list of diagnostic measures
}
\description{
Return diagnostics after on a mixture distribution object fittted with 'mx_metafit'
}
\examples{

breaks <- 0:28
counts <- c(36, 0, 2, 3, 4, 8, 9, 14, 10, 9, 3, 1, 1, 2,
            4, 8, 20, 45, 40, 54, 41, 22, 8, 3, 3, 0, 0,0)

observations <- unbin(breaks[-1], counts) # upper class boundaries

(comp <- mx_guess_components(observations, bw=2/3, mincut=0.9))
obj <- mxObj(comp, left="e")
ret <- mx_metafit(breaks, counts, obj)

summary(ret)       # general summary
cov2cor(vcov(ret)) # correlation matrix of parameters
diagnostics(ret)   # additional diagnostics

}
