% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_compare.R
\name{mx_compare}
\alias{mx_compare}
\title{Compare Fitted Mixture Distribution to Data}
\usage{
mx_compare(parms, breaks, counts, plot = FALSE, ...)
}
\arguments{
\item{parms}{paremeter list of a ..... mixture distribution}

\item{breaks}{upper class limits of the data}

\item{counts}{frequency of observations}

\item{plot}{boolean, whether to show a diagnostic plot}

\item{...}{additional arguments passed to \code{plot}}
}
\description{
Compare Fitted Mixture Distribution to Data
}
\examples{

}
