% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_components.R
\name{mx_guess_components}
\alias{mx_guess_components}
\title{Initial Guess of Univariate Mixture Components}
\usage{
mx_guess_components(x, bw = 1, minpeak = 0, mincut = 0.9, ...)
}
\arguments{
\item{x}{vector of (non-binned) observation data}

\item{bw}{bandwidth of kernel density, cf. \code{\link{density}}}

\item{minpeak}{minimum value of the total maximum which is regarded as peak}

\item{mincut}{minimum relative height of a pit compared to the lower of
the two neighbouring maxima at which these maxima are regarded as separate
peaks (default value is derived from golden section)}

\item{...}{further arguments passed to  \code{\link{density}}}
}
\value{
data frame with mean (\code{mean}), standard deviation \code{sd} and
  mixing proportions {L}
}
\description{
Heuristic method to guess number of components and their parameters
(mean, standard deviation and weight) from observation data.
}
\details{
The function guesses approximate start values of parameters
  that then need subsequent identification, e.g. with a maximum likelihood or
  EM (expectation maximization) method.
}
\examples{

set.seed(123)
x <- c(rnorm(20, 5, 1), rnorm(30, 10, 1), rnorm(50, 20, 2))
hist(x, breaks=20)

mx_guess_components(x)

}
\seealso{
\code{\link{peakwindow}}
}
