% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_parmlist_sort.R
\name{mx_parmlist_sort}
\alias{mx_parmlist_sort}
\alias{mx_parmlist_select}
\alias{mx_intermediate}
\alias{mx_wildtype}
\title{Sort and Select Distribution Components}
\usage{
mx_parmlist_sort(p)

mx_parmlist_select(p, N = 1, item = c("L", "mean", "sd"),
  ascending = TRUE)

mx_intermediate(p, as.vector = TRUE)

mx_wildtype(p, as.vector = TRUE)
}
\arguments{
\item{p}{mixture distribution parameter list}

\item{N}{number of components}

\item{item}{parameter name used for sorting and selection (e.g. 'L', 'mean', 'sd')}

\item{ascending}{ascending  (TRUE) ode descending (FALSE) order}

\item{as.vector}{if TRUE, result is returned as a vector, otherwise as list}

\item{...}{reserved for future extensions}
}
\value{

}
\description{
Sort and Select Distribution Components
}
\examples{

}
