% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_plot.R
\name{mx_plot}
\alias{mx_plot}
\title{Plot Components of Mixture Distribution}
\usage{
mx_plot(
  obj,
  breaks,
  counts,
  disc = 0,
  ecoff.prob = 0.01,
  main = "Subpopulations",
  length.out = 200,
  ...
)
}
\arguments{
\item{obj}{Object of class \code{\link{mxObj}}.}

\item{breaks}{class boundaries}

\item{counts}{frequency of observations}

\item{disc}{disc diameter (defaults to zero)}

\item{ecoff.prob}{probability threshold for the ecoff}

\item{main}{main title of the plot}

\item{length.out}{number of points used for plotting component densities}

\item{...}{other arguments passed to \code{\link{hist}}}
}
\value{
ecoff value (normal quantile)
}
\description{
Plot Components of Mixture Distribution
}
\examples{

breaks <- 0:28
counts <- c(36, 0, 2, 3, 4, 8, 9, 14, 10, 9, 3, 1, 1, 2,
            4, 8, 20, 45, 40, 54, 41, 22, 8, 3, 3, 0, 0,0)

observations <- unbin(breaks[-1], counts) # upper class boundaries

(comp <- mx_guess_components(observations, bw=2/3, mincut=0.9))

obj <- mxObj(comp, left="e")

obj2 <- mx_metafit(breaks, counts, obj)

mx_plot(obj2, breaks, counts, disc=5.5)

## simplification for fitted objects with save.data = TRUE
mx_plot(obj2, disc=5.5)

}
