% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxObj-class.R
\docType{class}
\name{mxMle-class}
\alias{mxMle-class}
\title{Mixture Distribution Object Containing an 'mle2' fit as Subobject}
\description{
The classes define ...
}
\section{Slots}{

\describe{
\item{\code{coef}}{list with the coefficients of the mixture distribution}

\item{\code{fit}}{complete results of the \code{mle2}-fit}

\item{\code{data}}{optionally, data used for the fit}
}}

