% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxObj-class.R
\docType{class}
\name{mxObj-class}
\alias{mxObj-class}
\title{Class of Univariate Mixture Distribution Components and its Parameters}
\description{
The classes define ...
}
\section{Slots}{

\describe{
\item{\code{coef}}{coefficients of the mixture distribution, specified by a
one-character ID of the distribution and the appropriate parameters}
}}

