% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abr_methods.R, R/mxObj-methods.R
\name{fit}
\alias{fit}
\alias{ecoff}
\alias{rsquared}
\alias{startpar}
\alias{models}
\alias{coef,mxObj-method}
\alias{pstart}
\alias{pstart,mxObj-method}
\alias{summary,mxMle-method}
\alias{vcov,mxMle-method}
\alias{AIC,mxMle-method}
\alias{logLik,mxMle-method}
\alias{get_data}
\alias{get_data,mxMle-method}
\alias{mxObj}
\title{Methods for 'mxObj' Univariate Mixture Objects}
\usage{
fit(object, ...)

ecoff(object, ...)

rsquared(object, ...)

startpar(object, ...)

models(object, ...)

\S4method{coef}{mxObj}(object, i = NULL, ...)

pstart(obj, ...)

\S4method{pstart}{mxObj}(obj, ...)

\S4method{summary}{mxMle}(object, ...)

\S4method{vcov}{mxMle}(object, ...)

\S4method{AIC}{mxMle}(object, ..., k = 2)

\S4method{logLik}{mxMle}(object, ...)

get_data(object, ...)

\S4method{get_data}{mxMle}(object, ...)

mxObj(x, type = NA, left = c("e", "n"))
}
\arguments{
\item{object}{an object of class \code{'mxObj'}}

\item{...}{optional arguments passed to the methods}

\item{i}{index of an individual distribution component}

\item{obj}{an object of class \code{'mxObj'}}

\item{k}{penalty parameter (nearly always left at its default value of 2)}

\item{x}{object (vector, list, data.frame) containing mixture distributions
in one of sereral known formats. (thpe-todo: document known formats)}

\item{type}{type of the mixture distribution, e.g. "n", "en", "nnn", etc.}

\item{left}{whether leftmost distribution is exponential ("e") or normal ("n")}
}
\description{
Constructor and accessor functions of
  univariate mixture distributions, composed of exponential, normal and gamma
  distributions.

The methods leverage methods from class 'mle2' of package 'bbmle'
}
\seealso{
\code{\link{mle2-class}}
}
