% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_oldclasses.R
\name{nls-class}
\alias{nls-class}
\alias{nls}
\title{Helper Classes}
\description{
Classes that import base functionality from other packages.
}
\seealso{
\code{\link{abr_baseclass}}, \code{\link{abr_ecoffinder-class}}
}
\concept{nls}
