% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmList.R
\name{parmList}
\alias{parmList}
\title{Convert Parameter Vector to List}
\usage{
parmList(p)
}
\arguments{
\item{p}{parameter vector of the following form: ..............}
}
\value{
list structure of mixture distribution parameters
}
\description{
deprecated! use as(vector, mxObj instead)
}
\details{
works for exponential, normal and gamma distributions
}
\examples{

}
