% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoffinder-plot-methods.R
\docType{methods}
\name{plot,abr_ecoffinder,ANY-method}
\alias{plot,abr_ecoffinder,ANY-method}
\title{Plot Fitted abrECOFFinder Object}
\usage{
\S4method{plot}{abr_ecoffinder,ANY}(x, cumulative = TRUE,
  fits = c("all", "best"), ...)
}
\arguments{
\item{x}{an object fitted by \code{\link{ecoffinder_nls}}}

\item{cumulative}{plot cumulative or density function}

\item{fits}{whether to plot all fits or only the best}

\item{\dots}{further parameters passed to \code{plot}}
}
\description{
Plot Fitted abrECOFFinder Object
}
\details{
: to do: documentation and examples
}
