% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_eucast.R
\name{reformat_eucast}
\alias{reformat_eucast}
\title{Reformat EUCAST crosstabe format into data base formats}
\usage{
reformat_eucast(data, type = c("ZD", "MIC"), melt = TRUE,
  unbin = FALSE)
}
\arguments{
\item{data}{data frame representing an EUCAST cross table or a result from
this function with "unbin=FALSE"}

\item{type}{either "ZD" or "MIC", the latter is not implemented yet}

\item{melt}{if \code{TRUE}, convert EUCAST crosstable to data base format}

\item{unbin}{expand binned data to single observtions}
}
\value{
reformatted data frame
}
\description{
Reformat EUCAST crosstabe format into data base formats
}
\details{
Reformatting of from crosstable to data base format is done with
function \code{melt} from package \pkg{reshape2}. In contrast,
\code{unbin} is a quite memory consuming workaround for R functions that do
not support a weight argument. Please consider to use weights and avoid
\code{unbin} whenever possible.
}
\note{
This function is under development, name and functionality may change.
}
