% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnpoints.R
\name{turnpoints}
\alias{turnpoints}
\title{Find turning points (peaks or pits)
Determine the number and the position of extrema (turning points, either peaks or pits) in a regular time series.}
\usage{
turnpoints(x)
}
\arguments{
\item{x}{vector or a time series}
}
\value{
An object of type 'abrTurnpoints' is returned, with:
\item{data}{ The dataset to which the calculation is done }
\item{n}{ The number of observations }
\item{points}{ The value of the points in the series, after elimination of ex-aequos }
\item{pos}{ The position of the points on the time scale in the series (including ex-aequos) }
\item{exaequos}{ Location of exaequos (1), or not (0) }
\item{nturns}{ Total number of tunring points in the whole time series }
\item{firstispeak}{ Is the first turning point a peak (\code{TRUE}), or not (\code{FALSE}) }
\item{peaks}{ Logical vector. Location of the peaks in the time series without ex-aequos }
\item{pits}{ Logical vector. Location of the pits in the time series without ex-aequos}
\item{tppos}{ Position of the turning points in the initial series (with ex-aequos) }
}
\description{
Find turning points (peaks or pits)
Determine the number and the position of extrema (turning points, either peaks or pits) in a regular time series.
}
\references{
Ibanez, F. (1982) Sur une nouvelle application de la theorie de
  l'information a la description des series chronologiques planctoniques.
  J. Exp. Mar. Biol. Ecol., 4:619-632

Grosjean, P. and Ibanez, F. (2014) pastecs: Package for Analysis
  of Space-Time Ecological Series. R package version 1.3-18.
  https://CRAN.R-project.org/package=pastecs
}
\author{
This is a stripped-down version taken from package \pkg{pastecs},
original authors are Frédéric Ibanez, Philippe Grosjean.
The original code and the derived version are licensed under the GPL (>=2).
}
