% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbin.R
\name{unbin}
\alias{unbin}
\title{Unbin Data Frame}
\usage{
unbin(x, freq)
}
\arguments{
\item{x}{a vector or data frame with binned data}

\item{freq}{vector of frequencies}
}
\value{
Vector or data frame with the same stucture as \code{x} with replicated
  elements or rows.
}
\description{
Unbin a data frame, i.e. replicate rows according to frequencies given in
a vector. The function is intended to reconstruct raw data from data
binned to classes.
}
\examples{

x <- data.frame(a=1:3, b=c("a", "b", "c"), freq=c(2, 5, 3))
unbin(x, x$freq)

zd <- 6:35

counts <- c(155, 0, 8, 12, 17, 35, 37, 66, 42, 39, 13, 4, 4, 8, 19, 36,
            80, 205, 188, 219, 170, 104, 32, 13, 12, 0, 3, 0, 0, 0)

observations <- unbin(zd, counts)

}
