% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimix-methods.R
\name{dunimix}
\alias{dunimix}
\alias{punimix}
\alias{runimix}
\alias{dunimix,numeric,list-method}
\alias{dunimix,numeric,mxObj-method}
\alias{punimix,numeric,list-method}
\alias{punimix,numeric,mxObj-method}
\alias{runimix,numeric,list-method}
\alias{runimix,numeric,mxObj-method}
\title{Univariate Mixtures of Exponential, Normal and Gamma Distributions}
\usage{
dunimix(x, obj, full.out, ...)

punimix(q, obj, full.out, ...)

runimix(n, obj, ...)

\S4method{dunimix}{numeric,list}(x, obj, full.out, ...)

\S4method{dunimix}{numeric,mxObj}(x, obj, full.out = FALSE, ...)

\S4method{punimix}{numeric,list}(q, obj, full.out = FALSE, ...)

\S4method{punimix}{numeric,mxObj}(q, obj, full.out, ...)

\S4method{runimix}{numeric,list}(n, obj, ...)

\S4method{runimix}{numeric,mxObj}(n, obj, ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{obj}{obkect of class \code{mxObj} with mixture distribution parameters}

\item{full.out}{return total density only or a matrix with all components}

\item{...}{reserved for future extensions}

\item{q}{vector of quantiles}

\item{n}{number of random values}
}
\value{
dunimix gives the density, punimix gives the distribution function,
  qunimix gives the quantile function, and runimix generates random deviates.
  The length of the result is determined by the legnt of x.
 Argument \code{full.out} determines whether only a vector with the totals of
   the distribution or a data frame with all components is returned.
}
\description{
Density, distribution function, quantile function and random generation for
  univariate mixture distributions, composed of exponential, normal and gamma
  distributions.
}
\note{
A quantile function \code{qunimix} is not yet implemented.
  Automatic coercion of arguments may still have bugs.
}
\examples{

}
