% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zd_main_components.R
\name{zd_main_components}
\alias{zd_main_components}
\title{Extract Main Components from Zone Diameter (ZD) Mixture Distribution}
\usage{
zd_main_components(p, np = 3, near_zero = 3)
}
\arguments{
\item{p}{data frame of distribution parameters (mean, sd, L) returned by
\code{\link{mx_guess_components}}}

\item{np}{number of desired components}

\item{near_zero}{heuristic upper boundary of the interval that is
considered as close to zero}
}
\value{

}
\description{
Extract np-1 main components and an additional first potentially
exponentially distributed component from a data frame of distribution
parameters with columns mean, sd, L.
}
\details{
The algorithm is pecifically intended to zone diameter (ZD)
  distributions where the leftmost (resistant) subpopulation is considered
  exponentially distributed, while the others (intermediate, wild-type)
  are considered approximately normal.
}
\examples{


set.seed(123)
x <- c(rexp(20, rate=1), rnorm(30, 10, 1), rnorm(10,15,.2), rnorm(50, 20, 2))
hist(x, breaks=20)
p <- mx_guess_components(x)

(ret <- zd_main_components(p, np=3))


}
