% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abr_densfunctions.R
\name{abr_density}
\alias{abr_density}
\alias{abr_cumdens}
\alias{abr_density.control}
\alias{abr_cumdensquant}
\alias{maxdens}
\title{Kernel Density Estimate for ZD and MIC data}
\usage{
abr_density(x, cutoff = 5.5, method = c("density", "evmix"),
  control = abr_density.control())

abr_cumdens(x, cutoff = 5.5, method = c("evmix", "density", "spline",
  "fmm"), control = abr_density.control())

abr_density.control(bw = 1, cut = 3,  n=512, bcmethod = "renorm", proper = TRUE)

abr_cumdensquant(x, p = 0.5, cutoff = 5.5, method = c("density",
  "evmix", "spline", "fmm"), control = abr_density.control())

maxdens(x, cutoff = 5.5, from = min(x), to = max(x),
  method = c("density", "evmix"), control = abr_density.control())
}
\arguments{
\item{x}{vector of antibiotic resistance observations, given as zone
diameter (zd) or log of minimum inhibitory concentration (mic).
Note that for \code{abr_density} x must contain raw observations,
not frequencies, cf. \code{\link{unbin}}, while for
abr_cumdens x and y contain binned data. If y is omitted, x must contain a
two-column matrix with value and frequency.}

\item{cutoff}{cutoff value (disc diameter) in case of zd data}

\item{method}{character which density method to use: \code{'density'},
\code{'evmix'} (function \code{pbckden}), \code{'spline'},
 or \code{'fmm'} (monotone spline)}

\item{control}{options passed to the density estimation methods, see \code{\link{abr_density.control}}}

\item{bw}{the smoothing bandwidth to be used,
see \code{\link{density}} for details.}

\item{cut}{by default, the values of from and to are cut bandwidths beyond
the extremes of the data, see \code{\link{density}} for details}

\item{n}{the number of equally spaced points at which the density is to be
estimated, see \code{\link{density}} for details}

\item{bcmethod}{character, boundary correction method}

\item{proper}{logical, whether density is renormalised to integrate to unity}

\item{p}{vector of probability quantiles}

\item{from, to}{search range for maximum (= mode) search}
}
\value{
data frame with
  \itemize{
    \item x zone diameter resp. mic value
    \item y density estimate
  }
}
\description{
Estimate kernel density for antibiotic resistance data using either standard
  density estimation with Gaussian kernels or boundary corrected density estimation
  with function \code{pbckden} from package \pkg{evmix}.
}
\details{
describe options ...
}
\examples{

}
\seealso{
\code{\link{density}}, \code{\link{dbckden}}
}
