% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_unimix.R
\name{fit_unimix}
\alias{fit_unimix}
\alias{fit_n}
\alias{fit_en}
\alias{fit_nn}
\alias{fit_enn}
\alias{fit_nnn}
\alias{fit_nnnn}
\alias{fit_ennn}
\alias{fit_ennnn}
\title{Fit of Normal and Exponential-Normal Mixtures to Binned Data}
\usage{
fit_unimix(breaks, counts, parms, type = c("n", "en", "enn", "ennn", "ennnn",
  "nn", "nnn", "nnnn"), ...)

fit_n(breaks, counts, parms, ...)

fit_en(breaks, counts, parms, ...)

fit_nn(breaks, counts, parms, ...)

fit_enn(breaks, counts, parms, ...)

fit_nnn(breaks, counts, parms, ...)

fit_nnnn(breaks, counts, parms, ...)

fit_ennn(breaks, counts, parms, ...)

fit_ennnn(breaks, counts, parms, ...)
}
\arguments{
\item{breaks}{upper class limits of the data}

\item{counts}{frequency of observations}

\item{parms}{list of initial parameters for the mixture}

\item{type}{of the mixture distribution, e.g. 'enn' for exponential-normal-normal}

\item{...}{additional arguments passed to \code{mle2}}
}
\value{

}
\description{
Fit mixture distributions to binned data with a maximum likelihood method,
  inspired by Venables and Ripley (2002)
}
\examples{

}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S.
   Fourth Edition. Springer, New York. ISBN 0-387-95457-0

   Bolker, Ben and R Development Core Team (2017) bbmle:
   Tools for General Maximum Likelihood Estimation. R package version 1.0.20.
   \url{https://CRAN.R-project.org/package=bbmle}
}
